requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
install==1.3.5
openai==1.6.1
pydantic==2.5.3
pydantic_core==2.14.6
pypdf==3.17.4
python-dotenv==1.0.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
